Move Splash damage attack to new class
